Move example apps to C-only API.
