+++ configure.ac
@@ -425,7 +425,9 @@ NL_RESTORE_WERROR
 AC_CONFIG_FILES([
 Makefile
 include/Makefile
+include/cli/Makefile
 include/crypto/Makefile
+include/ncp/Makefile
 include/platform/Makefile
 src/Makefile
 src/cli/Makefile
