+++ examples/posix/app/cli/Makefile.am
@@ -43,10 +43,11 @@ soc_LDADD                                                      =  \
     $(top_builddir)/src/cli/libopenthread-cli.a                   \
     $(top_builddir)/examples/posix/platform/libopenthread-posix.a \
     $(top_builddir)/third_party/mbedtls/libmbedcrypto.a           \
+    -lstdc++                                                         \
     $(NULL)
 
 soc_SOURCES                                                    = \
-    main.cpp                                                     \
+    main.c                                                       \
     $(NULL)
 
 include $(abs_top_nlbuild_autotools_dir)/automake/post.am
