+++ examples/posix/platform/serial.c
@@ -30,6 +30,7 @@
 #include <fcntl.h>
 #include <poll.h>
 #include <stdlib.h>
+#include <stdio.h>
 #include <termios.h>
 #include <unistd.h>
