+++ include/cli/cli-serial.h
@@ -28,75 +28,25 @@
 
 /**
  * @file
- *   This file contains definitions for a CLI server on the serial service.
+ * @brief
+ *  This file defines the top-level functions for the OpenThread CLI server.
  */
 
-#ifndef CLI_SERIAL_HPP_
-#define CLI_SERIAL_HPP_
+#ifndef CLI_SERIAL_H_
+#define CLI_SERIAL_H_
 
-#include <openthread-types.h>
-#include <cli/cli_server.hpp>
-#include <common/tasklet.hpp>
-
-namespace Thread {
-namespace Cli {
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 /**
- * This class implements the CLI server on top of the serial platform abstraction.
+ * Initialize the CLI serial server.
  *
  */
-class Serial: public Server
-{
-public:
-    Serial(void);
+void otCliSerialInit(void);
 
-    /**
-     * This method delivers raw characters to the client.
-     *
-     * @param[in]  aBuf        A pointer to a buffer.
-     * @param[in]  aBufLength  Number of bytes in the buffer.
-     *
-     * @returns The number of bytes placed in the output queue.
-     *
-     */
-    int Output(const char *aBuf, uint16_t aBufLength);
+#ifdef __cplusplus
+}  // extern "C"
+#endif
 
-    /**
-     * This method delivers formatted output to the client.
-     *
-     * @param[in]  aFmt  A pointer to the format string.
-     * @param[in]  ...   A variable list of arguments to format.
-     *
-     * @returns The number of bytes placed in the output queue.
-     *
-     */
-    int OutputFormat(const char *fmt, ...);
-
-    void ReceiveTask(const uint8_t *aBuf, uint16_t aBufLength);
-    void SendDoneTask(void);
-
-private:
-    enum
-    {
-        kRxBufferSize = 128,
-        kTxBufferSize = 512,
-        kMaxLineLength = 128,
-    };
-
-    ThreadError ProcessCommand(void);
-    void Send(void);
-
-    char mRxBuffer[kRxBufferSize];
-    uint16_t mRxLength;
-
-    char mTxBuffer[kTxBufferSize];
-    uint16_t mTxHead;
-    uint16_t mTxLength;
-
-    uint16_t mSendLength;
-};
-
-}  // namespace Cli
-}  // namespace Thread
-
-#endif  // CLI_SERIAL_HPP_
+#endif
